requirements.txt
Flask==0.12
Flask-SQLAlchemy==2.2
Flask-Script==2.0.5
Flask-Migrate==2.0.3
psycopg2==2.7
factory_boy
pytest
